[charts] Improve code reuse in `ChartsYAxis`
