+++ packages/x-charts/src/ChartsYAxis/ChartsSingleYAxis.tsx
@@ -1,41 +1,38 @@
 'use client';
 import * as React from 'react';
+import { useRtl } from '@mui/system/RtlProvider';
 import { useIsHydrated } from '../hooks/useIsHydrated';
-import { getStringSize } from '../internals/domUtils';
 import { useTicks } from '../hooks/useTicks';
 import { useDrawingArea } from '../hooks/useDrawingArea';
 import { ChartsYAxisProps } from '../models/axis';
 import { useChartContext } from '../context/ChartProvider';
 import { shortenLabels } from './shortenLabels';
-import { AXIS_LABEL_TICK_LABEL_GAP, TICK_LABEL_GAP, YAxisRoot } from './utilities';
-import { useAxisProps } from './useAxisProps';
+import { AXIS_LABEL_TICK_LABEL_GAP, TICK_LABEL_GAP } from './utilities';
+import { useAxisTicksProps } from './useAxisTicksProps';
+
+interface ChartsSingleYAxisProps extends ChartsYAxisProps {
+  axisLabelHeight: number;
+}
 
 /**
  * @ignore - internal component.
  */
-function ChartsSingleYAxis(inProps: ChartsYAxisProps) {
+function ChartsSingleYAxis(inProps: ChartsSingleYAxisProps) {
+  const { axisLabelHeight } = inProps;
   const {
     yScale,
     defaultizedProps,
     tickNumber,
     positionSign,
-    skipAxisRendering,
     classes,
-    Line,
     Tick,
     TickLabel,
-    Label,
     axisTickLabelProps,
-    axisLabelProps,
-    lineProps,
-    isRtl,
-  } = useAxisProps(inProps);
+  } = useAxisTicksProps(inProps);
+  const isRtl = useRtl();
 
   const {
-    position,
-    disableLine,
     disableTicks,
-    label,
     tickSize: tickSizeProp,
     valueFormatter,
     slotProps,
@@ -43,13 +40,10 @@ function ChartsSingleYAxis(inProps: ChartsYAxisProps) {
     tickLabelPlacement,
     tickInterval,
     tickLabelInterval,
-    sx,
-    offset,
     width: axisWidth,
   } = defaultizedProps;
 
   const drawingArea = useDrawingArea();
-  const { left, top, width, height } = drawingArea;
   const { instance } = useChartContext();
   const isHydrated = useIsHydrated();
 
@@ -65,23 +59,11 @@ function ChartsSingleYAxis(inProps: ChartsYAxisProps) {
     direction: 'y',
   });
 
-  // Skip axis rendering if no data is available
-  // - The domain is an empty array for band/point scales.
-  // - The domains contains Infinity for continuous scales.
-  // - The position is set to 'none'.
-  if (skipAxisRendering) {
-    return null;
-  }
-
-  const labelRefPoint = {
-    x: positionSign * axisWidth,
-    y: top + height / 2,
-  };
   /* If there's an axis title, the tick labels have less space to render  */
   const tickLabelsMaxWidth = Math.max(
     0,
     axisWidth -
-      (label ? getStringSize(label, axisLabelProps.style).height + AXIS_LABEL_TICK_LABEL_GAP : 0) -
+      (axisLabelHeight > 0 ? axisLabelHeight + AXIS_LABEL_TICK_LABEL_GAP : 0) -
       tickSize -
       TICK_LABEL_GAP,
   );
@@ -91,13 +73,7 @@ function ChartsSingleYAxis(inProps: ChartsYAxisProps) {
     : new Map(Array.from(yTicks).map((item) => [item, item.formattedValue]));
 
   return (
-    <YAxisRoot
-      transform={`translate(${position === 'right' ? left + width + offset : left - offset}, 0)`}
-      className={classes.root}
-      sx={sx}
-    >
-      {!disableLine && <Line y1={top} y2={top + height} className={classes.line} {...lineProps} />}
-
+    <React.Fragment>
       {yTicks.map((item, index) => {
         const { offset: tickOffset, labelOffset, value } = item;
         const xTickLabel = positionSign * (tickSize + TICK_LABEL_GAP);
@@ -138,12 +114,7 @@ function ChartsSingleYAxis(inProps: ChartsYAxisProps) {
           </g>
         );
       })}
-      {label && isHydrated && (
-        <g className={classes.label}>
-          <Label {...labelRefPoint} {...axisLabelProps} text={label} />
-        </g>
-      )}
-    </YAxisRoot>
+    </React.Fragment>
   );
 }
