+++ packages/x-charts/src/ChartsYAxis/ChartsYAxis.tsx
@@ -1,11 +1,26 @@
 'use client';
 import * as React from 'react';
 import PropTypes from 'prop-types';
+import useSlotProps from '@mui/utils/useSlotProps';
+import { useThemeProps, useTheme, styled } from '@mui/material/styles';
 import { warnOnce } from '@mui/x-internals/warning';
 import { ChartsYAxisProps } from '../models/axis';
-import { useYAxes } from '../hooks/useAxis';
 import { ChartsSingleYAxis } from './ChartsSingleYAxis';
 import { ChartsGroupedYAxis } from './ChartsGroupedYAxis';
+import { ChartsText, ChartsTextProps } from '../ChartsText';
+import { defaultProps, useUtilityClasses } from './utilities';
+import { isInfinity } from '../internals/isInfinity';
+import { useDrawingArea } from '../hooks/useDrawingArea';
+import { useIsHydrated } from '../hooks/useIsHydrated';
+import { useYAxes } from '../hooks';
+import { isBandScale } from '../internals/isBandScale';
+import { getStringSize } from '../internals/domUtils';
+import { AxisRoot } from '../internals/components/AxisSharedComponents';
+
+const YAxisRoot = styled(AxisRoot, {
+  name: 'MuiChartsYAxis',
+  slot: 'Root',
+})({});
 
 /**
  * Demos:
@@ -18,19 +33,108 @@ import { ChartsGroupedYAxis } from './ChartsGroupedYAxis';
  */
 function ChartsYAxis(inProps: ChartsYAxisProps) {
   const { yAxis, yAxisIds } = useYAxes();
+  const drawingArea = useDrawingArea();
+  const isHydrated = useIsHydrated();
+
+  const { scale: yScale, tickNumber, reverse, ...settings } = yAxis[inProps.axisId ?? yAxisIds[0]];
+  const themedProps = useThemeProps({ props: { ...settings, ...inProps }, name: 'MuiChartsYAxis' });
+  const defaultizedProps = { ...defaultProps, ...themedProps };
+  const classes = useUtilityClasses(defaultizedProps);
+
+  const { left, top, width, height } = drawingArea;
+  const theme = useTheme();
+
+  const {
+    position,
+    disableLine,
+    label,
+    labelStyle,
+    offset,
+    width: axisWidth,
+    sx,
+    slots,
+    slotProps,
+  } = defaultizedProps;
+
+  const positionSign = position === 'right' ? 1 : -1;
+  const Line = slots?.axisLine ?? 'line';
+  const Label = slots?.axisLabel ?? ChartsText;
+
+  const lineProps = useSlotProps({
+    elementType: Line,
+    externalSlotProps: slotProps?.axisLine,
+    additionalProps: {
+      strokeLinecap: 'square' as const,
+    },
+    ownerState: {},
+  });
+  const axisLabelProps = useSlotProps({
+    elementType: Label,
+    externalSlotProps: slotProps?.axisLabel,
+    additionalProps: {
+      style: {
+        ...theme.typography.body1,
+        lineHeight: 1,
+        fontSize: 14,
+        angle: positionSign * 90,
+        textAnchor: 'middle',
+        dominantBaseline: 'text-before-edge',
+        ...labelStyle,
+      },
+    } as Partial<ChartsTextProps>,
+    ownerState: {},
+  });
 
   const axis = yAxis[inProps.axisId ?? yAxisIds[0]];
-
   if (!axis) {
     warnOnce(`MUI X Charts: No axis found. The axisId "${inProps.axisId}" is probably invalid.`);
     return null;
   }
 
-  if ('groups' in axis && Array.isArray(axis.groups)) {
-    return <ChartsGroupedYAxis {...inProps} />;
+  const domain = yScale.domain();
+  const isScaleBand = isBandScale(yScale);
+  const skipAxisRendering =
+    (isScaleBand && domain.length === 0) ||
+    (!isScaleBand && domain.some(isInfinity)) ||
+    position === 'none';
+
+  // Skip axis rendering if no data is available
+  // - The domain is an empty array for band/point scales.
+  // - The domains contains Infinity for continuous scales.
+  // - The position is set to 'none'.
+  if (skipAxisRendering) {
+    return null;
   }
 
-  return <ChartsSingleYAxis {...inProps} />;
+  const labelRefPoint = {
+    x: positionSign * axisWidth,
+    y: top + height / 2,
+  };
+
+  const axisLabelHeight = label == null ? 0 : getStringSize(label, axisLabelProps.style).height;
+
+  const children =
+    'groups' in axis && Array.isArray(axis.groups) ? (
+      <ChartsGroupedYAxis {...inProps} />
+    ) : (
+      <ChartsSingleYAxis {...inProps} axisLabelHeight={axisLabelHeight} />
+    );
+
+  return (
+    <YAxisRoot
+      transform={`translate(${position === 'right' ? left + width + offset : left - offset}, 0)`}
+      className={classes.root}
+      sx={sx}
+    >
+      {!disableLine && <Line y1={top} y2={top + height} className={classes.line} {...lineProps} />}
+      {children}
+      {label && isHydrated && (
+        <g className={classes.label}>
+          <Label {...labelRefPoint} {...axisLabelProps} text={label} />
+        </g>
+      )}
+    </YAxisRoot>
+  );
 }
 
 ChartsYAxis.propTypes = {
