+++ packages/x-charts/src/ChartsYAxis/utilities.ts
@@ -1,8 +1,6 @@
 import composeClasses from '@mui/utils/composeClasses';
-import { styled } from '@mui/material/styles';
 import { AxisConfig, ChartsYAxisProps } from '../models/axis';
 import { getAxisUtilityClass } from '../ChartsAxis/axisClasses';
-import { AxisRoot } from '../internals/components/AxisSharedComponents';
 
 export const useUtilityClasses = (ownerState: AxisConfig<any, any, ChartsYAxisProps>) => {
   const { classes, position, id } = ownerState;
@@ -22,11 +20,6 @@ export const useUtilityClasses = (ownerState: AxisConfig<any, any, ChartsYAxisPr
 export const TICK_LABEL_GAP = 2;
 /* Gap between the axis label and tick labels. */
 export const AXIS_LABEL_TICK_LABEL_GAP = 2;
-
-export const YAxisRoot = styled(AxisRoot, {
-  name: 'MuiChartsYAxis',
-  slot: 'Root',
-})({});
 
 export const defaultProps = {
   disableLine: false,
